move ObstaclesContainer and VehiclesContainer classes files into Containers folder, same for their unit tests
