+++ MarsRover/Models/Plateaus/PlateauBase.cs
@@ -1,4 +1,5 @@
-﻿using MarsRover.Models.Positions;
+﻿using MarsRover.Models.Plateaus.Containers;
+using MarsRover.Models.Positions;
 
 namespace MarsRover.Models.Plateaus;
